+++ samples/chaincode/vltava/src/containerServices/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from "./containerServices";
+export * from "./orchestrator";
